Rewrite show action in more elixiry way (I think?)
Basic premise of this talk I was watching was to 1. use pattern matching
to avoid conditional logic inside function bodies, and 2. to use this
pattern to avoid error handling mixing in with business logic. We now
pattern match on errors and format types, so that the rendering methods
are all very straightforward, and just respond to one format/error
condition. Also, forgive the terrible naming, I was just trying to see
how the actual logic looks in this style.